Allow setting hbase-site.xml properties with environment variables.
Example: to use an external zookeeper cluster instead of the in-process zookeeper that comes along with hbase standalone, set HBASE_CONF_HBASE_ZOOKEEPER_QUORUM.

Any arbitrary property can be set with a variable with the prefix HBASE_CONF followed by an hbase property name, with '.' replaced with '_'. The property will be lowercased before insertion, so natural variable names are fine.